requirements.txt
# Core ML and NLP
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
sentencepiece>=0.1.99
sacrebleu>=2.3.1
sentence-transformers>=2.2.2
torch>=2.1.0

# Data Processing
pymupdf>=1.23.0
python-docx>=1.1.0
ebooklib>=0.18
# bleualign - install from GitHub: pip install git+https://github.com/rsennrich/Bleualign.git
# Or use simple alignment method (default) - bleualign is optional
nltk>=3.8.1

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.3
sqlalchemy>=2.0.23
alembic>=1.12.1

# Backend API
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# External APIs
openai>=1.3.0
google-generativeai>=0.3.0

# Utilities
redis>=5.0.1
celery>=5.3.4
numpy>=1.24.3
pandas>=2.1.3
scikit-learn>=1.3.2
faiss-cpu>=1.7.4

# Monitoring
prometheus-client>=0.19.0

